+++ gcovr/writer/html/__init__.py
@@ -24,14 +24,14 @@ import io
 from argparse import ArgumentTypeError
 from typing import Any, Callable, Dict, Optional, Union
 
-from ..version import __version__
-from ..utils import (
+from ...version import __version__
+from ...utils import (
     force_unix_separator,
     realpath,
     commonpath,
     open_text_for_writing,
 )
-from ..coverage import (
+from ...coverage import (
     CallCoverage,
     CovData,
     CoverageStat,
@@ -75,7 +75,7 @@ def templates():
     from jinja2 import Environment, PackageLoader
 
     return Environment(
-        loader=PackageLoader("gcovr"),
+        loader=PackageLoader("gcovr.writer.html"),
         autoescape=True,
         trim_blocks=True,
         lstrip_blocks=True,
@@ -210,7 +210,7 @@ class RootInfo:
         self.link_function_list = options.html_details or options.html_nested
         self.relative_anchors = options.relative_anchors
 
-        self.version = __version__
+        self.version = "master" if __version__.endswith("+master") else __version__
         self.head = options.html_title
         self.date = options.timestamp.isoformat(sep=" ", timespec="seconds")
         self.encoding = options.html_encoding